Fix issue with workspaces mounted as root
One student was not able to run git so far in the mounted folder inside
of the DevContainer.  It seems to be this issue:

https://github.com/microsoft/vscode-remote-release/issues/6683

Added a line to run a postCreateCommand to change the ownership of
the mounted /workspaces/ to be the vscode user.  This is not exactly
the solution described in the above link.  Still researching to see
if there is a better/correct way to have the folder mounted as the
vscode user.